requirements.txt
beautifulsoup4==4.13.3
boto3==1.37.24
feedparser==6.0.11
openai==1.70.0
python-dotenv==1.1.0
Requests==2.32.3